Refactored `lar::RegionTree`
